fix(data): put the list of cars and accessories in a separate file, separate data from logic
