+++ data/accessoriesData.js
@@ -1,0 +1,12 @@
+const accessories = [
+  { id: 1, name: "Floor mats ", price: 80 },
+  { id: 2, name: "Roof rack", price: 400 },
+  { id: 3, name: "Seat covers", price: 300 },
+  { id: 4, name: "Tow hitch", price: 100 },
+  { id: 5, name: "Sunshades", price: 50 },
+  { id: 6, name: "Parking sensors", price: 1200 },
+  { id: 7, name: "Seat-back organizer", price: 1200 },
+  { id: 8, name: "Rear-view camera", price: 1200 },
+];
+
+export default accessories;
