+++ data/carsData.js
@@ -1,0 +1,70 @@
+const cars = [
+  {
+    id: 1,
+    make: "Chrysler",
+    model: "PT Cruiser",
+    year: 2000,
+    power: "140 KM",
+    mileage: "189 000 km",
+    price: 5000,
+    image:
+      "https://cdn.pixabay.com/photo/2014/04/03/17/54/chrysler-312880_1280.jpg",
+  },
+  {
+    id: 2,
+    make: "Nissan",
+    model: "Juke",
+    year: 2023,
+    power: "209 KM",
+    mileage: "53 000 km",
+    price: 13000,
+    image:
+      "https://images.pexels.com/photos/19896348/pexels-photo-19896348/free-photo-of-parked-gray-nissan-juke-with-its-lights-on.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  },
+  {
+    id: 3,
+    make: "Audi",
+    model: "TT",
+    year: 2023,
+    power: "190 KM",
+    mileage: "79 000 km",
+    price: 17000,
+    image:
+      "https://images.unsplash.com/photo-1600549522690-7f575f21dbd7?q=80&w=2076&auto=format&fit=crop&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+  },
+  {
+    id: 4,
+    make: "Citroën",
+    model: "Ami",
+    year: 2019,
+    power: "290 KM",
+    mileage: "19 000 km",
+    price: 9000,
+    image:
+      "https://images.pexels.com/photos/21694179/pexels-photo-21694179/free-photo-of-citroen-ami-parked-in-city.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  },
+  {
+    id: 5,
+    make: "Toyota",
+    model: "Prius",
+    year: 2023,
+    power: "204 KM",
+    mileage: "189 000 km",
+    price: 18000,
+    image:
+      "https://images.unsplash.com/photo-1707050050343-352947150d4e?q=80&w=2226&auto=format&fit=crop&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+  },
+  {
+    id: 6,
+    make: "Fiat",
+    model: "Multipla Sportage",
+    year: 2008,
+    power: "168 KM",
+    mileage: "117 400 km",
+    price: 19000,
+    image:
+      "https://images.pexels.com/photos/17121437/pexels-photo-17121437/free-photo-of-racing-fiat-multipla-on-grid.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  },
+];
+
+export default cars;
