+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
     <link rel="stylesheet" href="styles/mainScreen.css" />
-    <script src="scripts/mainScreen.js" defer></script>
+    <script type="module" src="scripts/mainScreen.js"></script>
   </head>
 
   <body>
